Split road_graph_csr.rs (595 lines) into road_graph_csr/ subdirectory
Split into focused modules:
- graph.rs: CsrGraph struct and construction from RoadNetwork
- pathfinding.rs: A* pathfinding and BPR traffic-aware routing
- pathfinding_data.rs: async-friendly PathfindingData bundle
- tests.rs: all unit tests
- mod.rs: thin facade with re-exports
